Created multiple test files and CLI flags to run specific test files
extension.test.ts had two suites: suite 'Markdown and .treatments.yaml file detection' is in detection.test.ts, suite 'Diagnostics detection' is in diagnostics.test.ts
