+++ runExtensionTests.js
@@ -2,9 +2,14 @@ const path = require('path');
 const { runTests } = require('@vscode/test-electron');
 
 async function main() {
+  const workspacePath = path.resolve(__dirname, '.');
+  console.log("Workspace path: " + workspacePath + __dirname);
+
+  // Gets string from CLI with "npm test --TEST_FILES"
+  const [, , ...testFiles] = process.argv;
+  process.env.TEST_FILES = testFiles || '';
+
   try {
-    const workspacePath = path.resolve(__dirname, '.');
-    console.log("Workspace path: " + workspacePath + __dirname);
     await runTests({
       extensionDevelopmentPath: path.resolve(__dirname, '.'),
       extensionTestsPath: path.resolve(__dirname, './out/test/suite/index.js'),
